refactor: use stream insertion instead of fmtlib specialization
