+++ vowpalwabbit/core/include/vw/core/learner.h
@@ -28,8 +28,6 @@
 #undef VW_DEBUG_LOG
 #define VW_DEBUG_LOG vw_dbg::learner
 
-#include "vw/core/vw_string_view_fmt.h"
-
 #include "vw/common/future_compat.h"
 #include "vw/core/example.h"
 #include "vw/core/label_type.h"
