+++ vowpalwabbit/core/src/multilabel.cc
@@ -2,8 +2,6 @@
 // individual contributors. All rights reserved. Released under a BSD (revised)
 // license as described in the file LICENSE.
 
-#include "vw/core/vw_string_view_fmt.h"
-
 #include "vw/common/text_utils.h"
 #include "vw/core/example.h"
 #include "vw/core/model_utils.h"
@@ -13,8 +11,6 @@
 #include "vw/io/logger.h"
 
 #include <cfloat>
-// needed for printing ranges of objects (eg: all elements of a vector)
-#include <fmt/ranges.h>
 
 namespace MULTILABEL
 {
