+++ vowpalwabbit/core/src/simple_label.cc
@@ -2,23 +2,16 @@
 // individual contributors. All rights reserved. Released under a BSD (revised)
 // license as described in the file LICENSE.
 
-#include "vw/core/vw_string_view_fmt.h"
-
 #include "vw/common/string_view.h"
-#include "vw/core/accumulate.h"
-#include "vw/core/best_constant.h"
 #include "vw/core/cache.h"
 #include "vw/core/example.h"
 #include "vw/core/parse_primitives.h"
-#include "vw/core/vw.h"
 #include "vw/io/logger.h"
 
 #include <cfloat>
 #include <cmath>
 #include <cstdio>
 #include <cstring>
-// needed for printing ranges of objects (eg: all elements of a vector)
-#include <fmt/ranges.h>
 
 VW::simple_label::simple_label() { reset_to_default(); }
